Add "gitignores" scaffold.
This pulls gitignore content from Github's big
[list](https://github.com/github/gitignore). It also defines a shared variable
around languages (many of our scaffolds will need to know which languages the
project supports). After selecting languages, the question won't be asked
again (unless the user deletes the entry in their .yo-rc).

Perhaps contentious here: I selected four languages (Go, Node, Python, Ruby)
for us to claim to support. I'd argue that other languages are certainly
possible (and easily doable for gitignores, for example), but that we'll want
to only promise scaffolds for a subset. It's also necessary to prevent an
obnoxiously long list for the user to select from.